生物图像处理/2nd/requirements.txt
imageio==2.22.0
matplotlib==3.6.2
numpy==1.23.3
Pillow==9.4.0
